code/requirements.txt
nltk==3.8.1
numpy==1.25.2
peft==0.5.0
python_Levenshtein==0.25.1
torch==2.0.1
tqdm==4.65.0
transformers==4.35.2
